Test disabling w3c for all failing tests
