+++ vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/grid/contextmenu/ContextMenuGridIT.java
@@ -15,7 +15,6 @@
  */
 package com.vaadin.flow.component.grid.contextmenu;
 
-import java.net.URL;
 import java.util.List;
 import java.util.Locale;
 import java.util.concurrent.TimeUnit;
@@ -24,20 +23,15 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.remote.RemoteWebDriver;
 
+import com.vaadin.flow.component.AbstractNoW3c;
 import com.vaadin.flow.component.grid.testbench.GridElement;
-import com.vaadin.flow.testutil.AbstractComponentIT;
 import com.vaadin.flow.testutil.TestPath;
-import com.vaadin.testbench.Parameters;
-import com.vaadin.testbench.TestBench;
 import com.vaadin.testbench.TestBenchElement;
 
 @TestPath("context-menu-grid")
-public class ContextMenuGridIT extends AbstractComponentIT {
+public class ContextMenuGridIT extends AbstractNoW3c {
 
     private static final String OVERLAY_TAG = "vaadin-context-menu-overlay";
 
@@ -51,26 +45,7 @@ public class ContextMenuGridIT extends AbstractComponentIT {
     }
 
 
-    @Override
-    public void setup() throws Exception {
-        if (getRunOnHub(getClass()) != null
-                || Parameters.getHubHostname() != null) {
 
-            ChromeOptions options = new ChromeOptions();
-            options.addArguments(new String[]{"--headless", "--disable-gpu"});
-            options.setExperimentalOption("w3c", false);
-
-            options.merge(getDesiredCapabilities());
-            setDesiredCapabilities(getDesiredCapabilities());
-
-            WebDriver driver =  TestBench.createDriver(
-                    new RemoteWebDriver(new URL(getHubURL()), options));
-            setDriver(driver);
-        } else {
-            super.setup();
-        }
-    }
-    
     @Test
     public void contextClickOnRow_itemClickGetsTargetItem() {
         grid.getCell(56, 1).contextClick();
