+++ vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/grid/it/GridViewIT.java
@@ -15,6 +15,7 @@
  */
 package com.vaadin.flow.component.grid.it;
 
+import java.net.URL;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
@@ -30,8 +31,11 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
+import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.remote.RemoteWebDriver;
 
 import com.vaadin.flow.component.grid.ColumnTextAlign;
 import com.vaadin.flow.component.grid.testbench.GridColumnElement;
@@ -40,7 +44,10 @@ import com.vaadin.flow.component.grid.testbench.GridTHTDElement;
 import com.vaadin.flow.component.grid.testbench.GridTRElement;
 import com.vaadin.flow.data.provider.QuerySortOrder;
 import com.vaadin.flow.demo.TabbedComponentDemoTest;
+import com.vaadin.testbench.Parameters;
+import com.vaadin.testbench.TestBench;
 import com.vaadin.testbench.TestBenchElement;
+import com.vaadin.testbench.parallel.Browser;
 
 /**
  * Integration tests for the {@link GridView}.
@@ -48,6 +55,28 @@ import com.vaadin.testbench.TestBenchElement;
 public class GridViewIT extends TabbedComponentDemoTest {
 
     private static final String OVERLAY_TAG = "vaadin-context-menu-overlay";
+
+    @Override
+    public void setup() throws Exception {
+        if (Browser.CHROME == this.getRunLocallyBrowser() && (
+                getRunOnHub(getClass()) != null
+                        || Parameters.getHubHostname() != null)) {
+
+            ChromeOptions options = new ChromeOptions();
+            options.addArguments(
+                    new String[] { "--headless", "--disable-gpu" });
+            options.setExperimentalOption("w3c", false);
+
+            options.merge(getDesiredCapabilities());
+            setDesiredCapabilities(getDesiredCapabilities());
+
+            WebDriver driver = TestBench.createDriver(
+                    new RemoteWebDriver(new URL(getHubURL()), options));
+            setDriver(driver);
+        } else {
+            super.setup();
+        }
+    }
 
     @Test
     public void dataIsShown() throws InterruptedException {
