+++ vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/grid/it/HiddenColumnIT.java
@@ -22,12 +22,13 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
 
+import com.vaadin.flow.component.AbstractNoW3c;
 import com.vaadin.flow.component.grid.testbench.GridElement;
 import com.vaadin.flow.testutil.AbstractComponentIT;
 import com.vaadin.flow.testutil.TestPath;
 
 @TestPath("hidden-column")
-public class HiddenColumnIT extends AbstractComponentIT {
+public class HiddenColumnIT extends AbstractNoW3c {
 
     private GridElement grid;
