+++ vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/treegrid/it/AbstractTreeGridIT.java
@@ -4,11 +4,12 @@ import org.junit.Assert;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
+import com.vaadin.flow.component.AbstractNoW3c;
 import com.vaadin.flow.component.grid.testbench.GridColumnElement;
 import com.vaadin.flow.component.grid.testbench.TreeGridElement;
 import com.vaadin.flow.testutil.AbstractComponentIT;
 
-public abstract class AbstractTreeGridIT extends AbstractComponentIT {
+public abstract class AbstractTreeGridIT extends AbstractNoW3c {
 
     private TreeGridElement grid;
